Add test for React that ensures a ReviewList is rendered
