+++ spec/App.test.js
@@ -1,13 +1,12 @@
 import React from 'react';
-import renderer from 'react-test-renderer';
-import { mount } from 'enzyme';
+import { shallow, mount, render } from 'enzyme';
 
-import App, { Review } from '../client/src/index.jsx';
+import Review from '../client/src/components/app.jsx';
+import ReviewList from '../client/src/components/reviewList.jsx';
 
 describe('Review', () => {
-  test('snapshot renders', () => {
-    const component = renderer.create(<Review />);
-    let tree = component.toJSON();
-    expect(tree).toMatchSnapshot();
+  test('renders review list', () => {
+    const wrapper = shallow(<Review />);
+    expect(wrapper.find(ReviewList).length).toEqual(1);
   });
 });
